Moved uusd balance up in state.
